requirements.txt
# streamlit
numpy
pillow
fastapi
torch
transformers
matplotlib
opencv-python-headless
scikit-learn
imageio
